rm: unnecessary dependencies, refac: config
Removed flexibility dependency. Refactored `package.json` to set the
entry point (located in the `./src` directory), set `browserlist` within
`package.json` so that transpilation occurs for older browsers of a
specific target (`> 0.5%, last 2 versions, not dead`), and update the
`build` command and add a `start` command that use the newly set entry
point.